Shuffle dirs
- Move a bunch of code from the top-level pokesay -> src/pokesay
- add a test
