+++ src/pokesay/utils.go
@@ -1,0 +1,9 @@
+package pokesay
+
+import "log"
+
+func Check(e error) {
+	if e != nil {
+		log.Fatal(e)
+	}
+}
