+++ test/data/pokemon.json
@@ -1,0 +1,3 @@
+{"idx":"001","name":{"eng":"Bulbasaur","chs":"妙蛙种子","jpn":"フシギダネ","jpn_ro":"Fushigidane"},"slug":{"eng":"bulbasaur","jpn":"fushigidane","jpn_ro":"fushigidane"},"gen-7":{"forms":{"$":{"has_female":false,"has_right":false}}},"gen-8":{"forms":{"$":{"is_prev_gen_icon":true}}}}
+{"idx":"002","name":{"eng":"Ivysaur","chs":"妙蛙草","jpn":"フシギソウ","jpn_ro":"Fushigisou"},"slug":{"eng":"ivysaur","jpn":"fushigisou","jpn_ro":"fushigisou"},"gen-7":{"forms":{"$":{"has_female":false,"has_right":false}}},"gen-8":{"forms":{"$":{"is_prev_gen_icon":true}}}}
+{"idx":"003","name":{"eng":"Venusaur","chs":"妙蛙花","jpn":"フシギバナ","jpn_ro":"Fushigibana"},"slug":{"eng":"venusaur","jpn":"fushigibana","jpn_ro":"fushigibana"},"gen-7":{"forms":{"$":{"has_female":false,"has_right":false},"mega":{"has_female":false,"has_right":false}}},"gen-8":{"forms":{"$":{"is_prev_gen_icon":false},"gmax":{"is_prev_gen_icon":false},"mega":{"is_prev_gen_icon":true}}}}
